fix: Fixup #163 response code to -32602 when JSON polymorphics fail to deserialize due to missing discriminator (#170)
* fixup: Fix #163 response code to -32602 when JSON polymorphics fail to deserialize due to missing discriminator

* Remove usage of [JsonPolymorphic] in favor of custom "converter"

It's not really a "converter" per se, but we want to control the errors
that happen due to incorrect format related to polymorphism (e.g.
missing or incorrect 'kind' value), and this is the only way to do it
without requiring consumers to de/serialize with not-native STJ constructs

* Remove unused MessageRoleConverter

* Fixes up de/serialization of A2AEvent chain

* Revert MessageJsonValue change in favor of new custom JSON
converter
* Add readonly Kind value to A2AEvent
* Add enum set for 'kind' discriminator values as this collection is
defined by the spec and must be handled by the code. Enum allows for
statis analysis to throw when all values in the enum aren't handled by
the event converter
* Add new A2AJsonConverter that can be used any time we want JSON
de/ser to throw A2AException with InvalidRequest
* Applied A2AJsonConverter to MessageSendParams
* Added UT to test for unknown/bad 'kind' value

* copilot suggestion

* adding caching for the A2AJsonConverters "safe" options for perf

* fixing var names

---------

Co-authored-by: Brandon🌩️H <brandonh-msft@users.noreply.github.com>